[upd]完成了近战塔和远程塔的放置机制, 下一步完成交战机制

+++ TowerDefense/enemy.h
@@ -13,7 +13,7 @@ class Enemy: public QObject, public QGraphicsPixmapItem
     int hp;
     int maxHp;
     int atk;
-    int atkRange;
+    float atkRange;
     float atkSpeed;
     float moveSpeed;
     int pathIdx;//记录属于哪一条路径
